+++ TicketManger/Ticket.cpp
@@ -2,12 +2,12 @@
 #include<memory>
 #include"Ticket.h"
 using namespace std;
-void Ticket::displayticket()
+void Ticket::DisplayTicket()
 {
-    cout<<"Source : "<<getsource()<<endl;
-    cout<<"Destination : "<<getdestination()<<endl;
+    cout<<"Source : "<<getSource()<<endl;
+    cout<<"Destination : "<<getDestination()<<endl;
     cout<<"PNR : "<<genratePNR()<<endl;
-    cout<<"TYPE :"<<getclasstype()<<endl;
-    cout<<"Ticket staus : "<<getstatus()<<endl;
-    cout<<"Seat NO. : "<<getseatno()<<endl;
+    cout<<"TYPE :"<<getClasstype()<<endl;
+    cout<<"Ticket staus : "<<getStatus()<<endl;
+    cout<<"Seat NO. : "<<getSeatno()<<endl;
 }
