+++ TicketManger/Ticket.h
@@ -2,27 +2,21 @@
 #include<memory>
 
 using namespace std;
-class Ticket
+class Ticket:public std::enable_shared_from_this<Ticket>
 {
     private:
-    Ticket();
+        Ticket();
     
     public:
-    std::enable_shared_from_this<Ticket>
-    {
-        public:
         typedef std::shared_ptr<Ticket> ptr;
-    }
-    public:
-    
-    void genrateticket();
-    string getname();
-    string getsource();
-    string getdestination();
-    int genratePNR();
-    string getclasstype();
-    string getstatus();
-    int getseatno();
-    void displayticket();
+        void genrateTicket();
+        string getName();
+        string getSource();
+        string getDestination();
+        int genratePNR();
+        string getClasstype();
+        string getStatus();
+        int getSeatno();
+        void DisplayTicket();
     
 };
