+++ TicketManger/Ticket_IF.h
@@ -1,0 +1,9 @@
+#include<iostream>
+#include<memory>
+using namespace std;
+class TicketIF
+{
+    public:
+    typedef std::shared_ptr<TicketIF> ptr;
+    
+};
