+++ TicketManger/Ticket_manager_IF.h
@@ -1,0 +1,15 @@
+
+#include<iostream>
+#include<memory>
+using namespace std;
+class Ticket_manager_IF
+{
+public:
+    typedef std::shared_ptr<Ticket_manager_IF> ptr;
+    virtual void book_ticket()=0;
+    virtual void cancel_ticket()=0;
+    virtual void prepared_chart()=0;
+
+    
+};
+
