--- TicketmangerIF/Ticket_IF.h
@@ -1,9 +1,0 @@
-#include<iostream>
-#include<memory>
-using namespace std;
-class TicketIF
-{
-    public:
-    typedef std::shared_ptr<TicketIF> ptr;
-    
-};
