--- TicketmangerIF/Ticket_manager_IF.cpp
@@ -1,8 +1,0 @@
-class TicketmanagerIF
-{
-    public:
-    virtual void book_ticket()=0;
-    virtual void cancel_ticket()=0;
-    virtual void prepared_chart()=0;
-
-};
